pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lerim"
version = "0.1.70"
description = "Continual learning layer for coding agents and software projects."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSL-1.1" }
dependencies = [
    "pydantic==2.12.5",
    "eval_type_backport==0.3.1; python_version < '3.13'",
    "tomli>=2.0; python_version < '3.11'",
    "python-dotenv==1.2.2",
    "PyYAML==6.0.3",
    "loguru==0.7.3",
    "dspy==3.1.3",
    "python-frontmatter==1.1.0",
    "openrouter==0.7.11",
    "logfire[dspy]==4.30.0",
    "litellm==1.81.13",
    "rich==14.3.3",
]

[project.urls]
Homepage = "https://lerim.dev"
Documentation = "https://docs.lerim.dev"
Repository = "https://github.com/lerim-dev/lerim-cli"
Changelog = "https://github.com/lerim-dev/lerim-cli/blob/main/CHANGELOG.md"
Issues = "https://github.com/lerim-dev/lerim-cli/issues"

[project.optional-dependencies]
mlx = [
    "vllm-mlx>=0.2.6",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-xdist>=3.5",
    "pytest-timeout>=2.4",
]
lint = [
    "ruff>=0.15.0",
    "vulture>=2.14",
]
docs = [
    "mkdocs-material>=9.6",
    "mkdocstrings[python]>=0.27",
]

[project.scripts]
lerim = "lerim.server.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["lerim*"]

[tool.setuptools.package-data]
lerim = ["config/default.toml", "skills/*.md"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests/unit", "tests/smoke", "tests/integration", "tests/e2e"]
markers = [
    "integration: Tests that hit real external services",
    "e2e: End-to-end tests that exercise full flows",
    "smoke: Quick LLM sanity checks (requires ollama)",
    "agent: Agent SDK or CLI integration tests",
    "llm: Non-agentic LLM integration tests",
    "openrouter: OpenRouter provider tests",
    "openai: OpenAI provider tests",
    "zai: ZAI provider tests",
    "kimi: Kimi provider tests",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
extend-select = ["B006"]
ignore = ["E501"]

[tool.vulture]
paths = ["src/"]
min_confidence = 100

[dependency-groups]
dev = [
    "jupyterlab>=4.5.5",
    "matplotlib>=3.10.8",
]
